python/requirements.txt
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
gevent==24.11.1
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
kombu==5.4.2
prompt_toolkit==3.0.50
protobuf==5.29.3
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
